+++ ElvUI_EltreumUI/Core/VersionCheck.lua
@@ -19,7 +19,7 @@ end
 
 function ElvUI_EltreumUI:ElvUIVersionCheck()
 	if ElvUI_EltreumUI.Retail then
-		if E.version < 12.31 then
+		if E.version < 12.32 then
 			ElvUI_EltreumUI:Print("Your ElvUI version is out of date, please update to avoid issues!")
 		end
 	end