+++ Changelog.md
@@ -6,6 +6,7 @@ ___
 Added Class Icons by Releaf\
 Added a second Details profile that includes class icons by Releaf
 Added an option to switch icon types on character profile
+Added custom dead and offline indicators in Retail using Releaf's textures
 Updated chat install according to ElvUI changes
 Fixed switching Dark Mode > Light Mode > Dark mode\
 Fixed cooldown being enabled when cursor was disabled